+++ src/main/java/kg/attractor/job_search_java_25/model/Vacancy.java
@@ -6,6 +6,7 @@ import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -31,7 +32,7 @@ public class Vacancy {
 
 
     @Column(name = "salary")
-    private Float salary;
+    private BigDecimal salary;
 
     @Column(name = "exp_from")
     private Integer expFrom;
